requirements.txt
Pillow
youtube-dl
lepl
python-telegram-bot==12.0.0b1
pydub